[#11059] refactor(iceberg-rest): Consolidate registerTable overwrite into MetastoreRegisterTableUtils
Replace RegisterTableOverwrite and per-backend PointerSwap classes with shared
MetastoreRegisterTableUtils and backend-specific overwriteMetadataLocation hooks.
